refactor(middleware): rename middleware files and update project references
