+++ Raftel.sln
@@ -38,6 +38,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raftel.Api.Client.UnitTests
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raftel.Application.IntegrationTests", "tests\Raftel.Application.IntegrationTests\Raftel.Application.IntegrationTests.csproj", "{C8DD88EE-0C79-48BB-BEA8-3447D003C89F}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "demo", "demo", "{0B37B4C7-E65F-4AF4-81E9-994579E41B4D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -53,13 +55,13 @@ Global
 		{9CD06F2C-90DD-4017-8D7A-C9DB63200F06} = {2628C66E-9787-4BDA-A4AF-EAA96AFE0574}
 		{1A6CB3E5-2CC7-451D-8DE7-D9E6C1C1661D} = {37FB5593-6445-48AD-9D92-52167A3B526D}
 		{DEF38B61-773C-4281-8A74-D1C9845D76F3} = {BB519B61-3B65-4069-A09F-01001DC650B9}
-		{EF8AB585-E523-4A0C-B8B1-E7DF54EE2FBA} = {BB519B61-3B65-4069-A09F-01001DC650B9}
 		{D77D9265-F8D6-442C-9B96-FE0040A42658} = {BB519B61-3B65-4069-A09F-01001DC650B9}
 		{BCC6B3AA-82CA-4099-8264-780742A04525} = {37FB5593-6445-48AD-9D92-52167A3B526D}
 		{749BF784-35C8-456F-A69D-7E8991FC17F9} = {BB519B61-3B65-4069-A09F-01001DC650B9}
 		{DD2DB4E4-C5FA-425F-A3C0-3EFDDBA78CF1} = {37FB5593-6445-48AD-9D92-52167A3B526D}
 		{A018F9FB-E01D-44CE-98C5-51EF922C06FB} = {BB519B61-3B65-4069-A09F-01001DC650B9}
 		{C8DD88EE-0C79-48BB-BEA8-3447D003C89F} = {BB519B61-3B65-4069-A09F-01001DC650B9}
+		{EF8AB585-E523-4A0C-B8B1-E7DF54EE2FBA} = {0B37B4C7-E65F-4AF4-81E9-994579E41B4D}
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{2EFFCFF4-3D9B-4681-9B4F-64E9F59545BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
