+++ src/Raftel.Application/Middlewares/IGlobalMiddleware.cs
