+++ src/Raftel.Application/Middlewares/IQueryMiddleware.cs
