+++ src/Raftel.Application/Middlewares/MiddlewareRegistry.cs
