+++ tests/Raftel.Infrastructure.Tests/Data/UnitOfWorkTests.cs
@@ -56,23 +56,4 @@ public class UnitOfWorkTests : IntegrationTestBase
             pirate.ShouldBe(loaded);
         });
     }
-
-    [Fact]
-    public async Task CommitAsync_ShouldPersistData_UsingMiddleware()
-    {
-        await ExecuteScopedAsync(async sp =>
-        {
-            var commandDispatcher = sp.GetService<ICommandDispatcher>();
-
-            var command = new CreatePirateCommand("Ace", 9514361);
-            var result = await commandDispatcher.DispatchAsync(command);
-            result.IsSuccess.ShouldBeTrue();
-
-            var repository = sp.GetRequiredService<IPirateRepository>();
-
-            var loaded = await repository.ListAllAsync();
-            loaded.ShouldHaveSingleItem();
-            loaded.ShouldContain(_ => _.Name == command.Name && _.Bounty == command.Bounty);
-        });
-    }
 }
