+++ tests/Raftel.Infrastructure.Tests/Raftel.Infrastructure.Tests.csproj
@@ -33,7 +33,6 @@
 
     <ItemGroup>
         <ProjectReference Include="..\..\src\Raftel.Infrastructure\Raftel.Infrastructure.csproj"/>
-        <ProjectReference Include="..\Raftel.Application.UnitTests\Raftel.Application.UnitTests.csproj" />
         <ProjectReference Include="..\Raftel.Domain.Tests\Raftel.Domain.Tests.csproj" />
     </ItemGroup>
